+++ BetterChests/CHANGELOG.md
@@ -15,6 +15,7 @@
 * Disallow any Chest from being automatically stashed into another chest.
 * Craft From Chest will only lock chests when crafting items.
 * Allow stacking of empty chests.
+* When opening a held chest, gray out that chest in inventory.
 
 ## 2.5.5 (August 27, 2022)